requirements.txt
# Core Dependencies
eel>=0.16.0
pyaudio>=0.2.11
pyautogui>=0.9.54
pvporcupine>=2.2.0
hugchat>=0.0.47
playsound>=1.3.0

# Enhanced Speech Recognition
SpeechRecognition>=3.10.0
webrtcvad>=2.0.10
numpy>=1.24.0

# Natural Language Processing
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0

# Face Recognition
face-recognition>=1.3.0
opencv-python>=4.8.0

# Security
cryptography>=41.0.0

# Cross-Platform Support
pygame>=2.5.0

# Modern TTS Engines
pyttsx3>=2.90
azure-cognitiveservices-speech>=1.34.0
TTS>=0.22.0

# Additional Features
pywhatkit>=5.4.3
requests>=2.31.0
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced Features
# openai>=1.3.0  # For ChatGPT integration
# google-cloud-speech>=2.21.0  # For Google Speech API
# boto3>=1.34.0  # For AWS Polly TTS 